feat(order_repository,firebase): support order with **items** CRUD
Note intricacies of adapting the order's subcollection of items to the local UI frontend model array
of items. Create new order with a subcollection new items (batch write). Read order and async get
the subcollection of items to copyWith into the local items array. Batch delete the subcollection
when we are deleting the item (we should alternatively use a Google Cloud Function).