refactor(boundary_data_transducer): extract out primed accumulator - prune iterator fn combine bloat
May be slightly gold plating in a way that is not very comprehensible. Weigh up whether to have the
transform function passed in, lots of parameter chaining code smell. Strangeness passing in the pure
form of ConstructorTransform vs the exception-handled try optional form.

On code review: we could get away with moving the tryParse import into
the big implementation body as we do not want to trust a parent callee
to make sure that they have created an exception-handled version?

Wait nvm. We already wrap it (the pure constructor) in the tryParseDataTransform function and expect API users to only access the `constructValidObjectsFromData` function.

Make sure the T vs List<T> is understandable... with all the generics overloading use of T
when T is List<T> depending on where you (the dev) start(s) reading from.

bloated commit message context.