test(boundary_data_transducer): check success/faliure criteria, graceful mix success conversions
Check for happiest path valid conversions, check invalid conversions all emptied, check mix retained
valid conversions only. TDD we found that the Exception handler catch did not appropriately handle
the case of missing keys with our example class/constructor. This was due to the map field not
checking null (and we do not want it to silently create the object when not all fields are valid,
anyway). Expected an exception to be raised; however, Dart throws a _CastError that we now handle
with TypeError catching.

The alternative would be to expect the developer to balloon out every
class parsing boilerplate (manual handcoding at the moment as template
are not all updated!) maybe double the number of lines to read/maintain
depending on the types of nested objects and the onus of adding null checking and argument exception throws to every map[''] field subscript access operation before use...

Potential use case for code generation, json serialisation possibility?
Currently using VS Code data class generator extensions + handcoded
amendments to linting and convention