refactor(repositories/models,drug,drug_dose,product): add exception, debug convert, self-explanatory
More self-explanatory code for non-happy path when parts of map are unexpected. Nulls. Also note
supply chain issue  type conversion error cast. Should have ocs link as strictly int. Previously had
erroneus data of the id number in a string field instead - causing complete breakage of the item. We
do not accommodate this further as we want the data to be clean from the start to prevent
inconsistency unmaintainability in the frontend logic domain